3518 - add subSection.props.hints meta and handle in Title
